pyproject.toml
[tool.poetry]
name = "to-do-list"
version = "0.1.0"
description = ""
authors = ["alexeydoronin <neonixys@mail.ru>"]
readme = "README.md"
packages = [{include = "to_do_list" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
envparse = "^0.2.0"
env = "^0.1.0"
gunicorn = "^20.1.0"
pre-commit = "^3.2.2"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.1"
social-auth-app-django = "^5.2.0"
drf-spectacular = "^0.26.1"
psycopg2-binary = "^2.9.6"
django-filter = "^23.1"




[tool.poetry.group.dev.dependencies]
ansible-vault-win = "^1.0.0"
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
